Complete gradient_descent_reg test

+++ mlpractice/tests/gradient_descent/test_gradient_descent.py
@@ -127,7 +127,7 @@ def test_random(descent=GradientDescent, iterations=1):
         np.random.seed(214)
         for test in range(iterations):
             inp_iteration = int(np.random.random() * 100) + 1
-            inp_lambda = int(np.random.random() * 10)
+            inp_lambda = np.random.random() * 10
             samples_number = int(np.random.random() * 100) + 1
             weight_length = int(np.random.random() * 9) + 1
             inp_w0 = 20 * np.random.ranf((weight_length, 1)) - 10
